+++ packages/ui/src/components/textarea.tsx
@@ -3,8 +3,7 @@ import * as React from "react";
 import { controlFieldBaseClasses } from "../lib/control-styles";
 import { cn } from "../lib/utils";
 
-export interface TextareaProps
-  extends React.ComponentProps<"textarea"> {}
+export type TextareaProps = React.ComponentProps<"textarea">;
 
 const Textarea = React.forwardRef<HTMLTextAreaElement, TextareaProps>(
   ({ className, ...props }, ref) => {